requirements.txt
psycopg2-binary
python-dotenv
polars[all]
requests